Update scriptExperiment.js
Working on a way to generate a list that just shows guessed letters that are incorrect. When I list out the guessed letters I am getting duplicates and have yet to determine why. I have used the new Set function to eliminate that  problem for now. I am experimenting with .splice, or some combination of .pop() or .push() to get the list of incorrect guessed letters.